decorator.py
import jwt
from flask import request
from flask_restx import abort
from constants import JWT_SECRET, JWT_ALGORITHM


# Декоратор для проверки авторизации и прав администратора
def admin_required(func):
    def wrapper(*args, **kwargs):
        if 'Authorization' not in request.headers:
            abort(401)
        data = request.headers['Authorization']
        token = data.split("Bearer ")[-1]
        role = None
        try:
            user = jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
            role = user.get('role', 'user')

        except Exception as e:
            print("JWT Decode Exсeption", e)
            abort(401)

        if role != 'admin':
            abort(403)

        return func(*args, **kwargs)

    return wrapper


# Декоратор для проверки авторизации
def auth_required(func):
    def wrapper(*args, **kwargs):
        if 'Authorization' not in request.headers:
            abort(401)
        data = request.headers['Authorization']
        token = data.split("Bearer ")[-1]
        try:
            jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
        except Exception as e:
            print("JWT Decode Exсeption", e)
            abort(401)

        return func(*args, **kwargs)

    return wrapper
