requirements.txt
aniso8601==9.0.1
attrs==21.2.0
click==8.0.3
flake8==4.0.1
Flask==2.1.3
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
MarkupSafe==2.0.1
marshmallow==3.13.0
mccabe==0.6.1
prettytable==2.2.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyrsistent==0.18.0
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.25
wcwidth==0.2.5
werkzeug==2.1.2
PyJWT==2.5.0
